backend/requirements.txt
langgraph
langchain-google-genai
langchain-core
python-dotenv
pydantic
ddgs
firecrawl-py
